organizing allocation methods

+++ src/core/memory/buddy-alloc.h
@@ -8,8 +8,8 @@ namespace leep
     const uint32_t kHeaderSize = 8;
     const uint32_t kMinAllocExp = 4;
     const uint32_t kMaxAllocExp = 30;
-    const uint32_t kMinAlloc = 1 << kMinAllocExp;
-    const uint32_t kMaxAlloc = 1 << kMaxAllocExp;
+    const size_t kMinAlloc = 1 << kMinAllocExp;
+    const size_t kMaxAlloc = 1 << kMaxAllocExp;
     const uint32_t kBlockCount = kMaxAllocExp - kMinAllocExp + 1;
 
     struct CList
@@ -36,10 +36,10 @@ namespace leep
         uint32_t getNode(int8_t *ptr, uint32_t block);
         bool parentSplit(uint32_t index);
         void flipParentSplit(uint32_t index);
-        uint32_t adequateBlock(uint32_t request_size);
+        uint32_t adequateBlock(size_t request_size);
         void lowerBlockLimit(uint32_t block);
-        void *alloc(uint32_t size);
-        template<typename T> T *allocT(uint32_t count = 1);
+        void *alloc(size_t size);
+        template<typename T> T *allocT(size_t count = 1);
         void free(void *ptr);
         BuddyAlloc();
         void init();
@@ -54,10 +54,10 @@ namespace leep
     };
 
     template<typename T>
-    T *BuddyAlloc::allocT(uint32_t count)
+    T *BuddyAlloc::allocT(size_t count)
     {
         T* mem = (T*)alloc(sizeof(T) * count);
-        for (uint32_t i = 0; i < count; ++i)
+        for (size_t i = 0; i < count; ++i)
         {
             new(mem + i) T();
         }
@@ -66,4 +66,3 @@ namespace leep
 }
 
 #endif  // __LEEP_CORE_MEMORY_BUDDY_ALLOCATOR_H__
-
